fix: make wrapper launchable
