+++ wrapper/interfaces/__init__.py
@@ -1,1 +1,4 @@
-from .interfaces import *
+from .clients import IRomApi
+from .modules import IKernelBuilder, IAssetsCollector
+from .engines import IContainerEngine, IDockerEngine
+from .commands import IBundleCommand
